requirements_test.txt
pytest==8.2.0
pytest-django==4.8.0
coverage==7.5.1
pytest-cov==5.0.0

# selenium==4.21.0